textparse: Call `Compact` on histograms after parsing
This happens for both `Histogram` and `FloatHistogram`. In this way,
if targets decide to optimize the exposition size by merging spans
with empty buckets in between, we still get a normalized results. It
will also normalize away any valid but weird representations like
empty spans, spans with offset zero, and empty buckets at the start or
end of a span.

Signed-off-by: beorn7 <beorn@grafana.com>